The Unification of Scenes *Warning* Big Update
I combined the Menu scene with the Game scene so that way we dont have the lag between transitions. Now whole app has camera background. Made the menu fully AR. Kept the round select submenu. In the main menu: 2 buttons, one for going to round selection menu and one for starting the game. Looking at the right you can see the options panel, which contains the summary of the game and the sound button ready for future additions. On the left side you can see the credits panel (I lost the credits script and couldnt restore it from git so someone who has it saven could at it). Looking at the bottom you can see.. nothing actually, but there is an invisible buttton that if pressed activates and deactivates the clouds secret on the "sky". I Unified the scripts for the camera usage so there is now only one running for the physical camera and one for the game camera. Deleted what wasnt needed from the files and from inside the Gamemanager script. Another big change is in handling the menu, there is no bullet or suchin the menu so i used the onmouseenter/exit/down unity methods to make the menu work so i had to lock the cursor in the menu. wherever the middle screen points is highlighted and wherever you press triggers the button currently on the center of the screen. Hitting play on unity editor makes the mouse invisible and locks it on screen, holding down the right mouse button u can rotate the ingame camera and left clicks triggers buttons. To gain access outside the game inspector simply hit the escape button on your keyboard. Stasrting the game unlocks the cursor cause ingame you can press other things except the screen center so you have to see the mouse. This difference is only visible in the unity editor, in the android app u dont feel the defference cause the mouse is always invisible and goes wherever you tap on the screen. While on the mainmenu in the android app hitting the back button on your device exits the app, thous there is no need for an exit button because in unity editor u can just stop the simulation. Scripts: made ascript for handling the buttons on the menu main except for the mute sound button which is united with the in game same use button thous it requires its own script handling it. The physical camera input is purel;y handled by the CameraManager script which is a unification of the previous mainmenu and ingame scripts along with the CameraRotaron script, as for the fittoscreen my latest method is used so there is no need for the fitscript. Im not going to say anything about the secred button and the secred in the clouds because it wouldnt be a secret, so if you want to know what it is you have to figure it out on your own. Didnt alter anything from the actual game except that when you lose and tap the scren instead of restarting the game it takes you back to the maim menu. If i used something i shouldnt have, used horrific scripting language and methods, altered something i shouldnt have, removed something that i shouldnt or in general terms did something wrong feel free to excuse me and scold me. Whatever it is that i did wrong thanks to GitHub we can return that or the whole thing back to how it was. This concludes my commit.

Cincerely,
Ioannis Tantaoui (Tnt)