Create scene for testing persistent camera
For the new menu we need to make the feedback of the device's camera
to persist through the scene changing and not be destroyed. This commit
does not need to be merged into master and is intended for testing
purposes only.

The first scene contains the background image and the controller.
The second scene contains only a message that when seen, it means that
the test was successful.

Signed-off-by: AnimaRain <ioannis_spyr@yahoo.gr>