Remove duplicate prefabs
The prefabs in Resources were hardlinks to the prefabs in Prefabs directory.
Despite that, Unity sees both the original file and the hardlink as two different
files and that messes up the metadata.
Having all the prefabs under the some directory for easier book-keeping is not
that essential, so the prefabs matching the prefabs in Resources can be removed.

Signed-off-by: AnimaRain <ioannis_spyr@yahoo.gr>