Fix enemy bullets exploding and giving points without being shot by player
Whenever a spawnable gets destroyed – shot by player, or destroyed when the scene closes,
or destroyed when colliding with player – it explodes and gives player points. That
should only happen when they get destroyed by a player's bullet.

EnemyBullets were not intended to have the explosion effect, so it got removed alltogether.

Signed-off-by: AnimaRain <ioannis_spyr@yahoo.gr>