Fix mishandling spawners between rounds
Spawners where not always being stashed correctly. This caused
errors with the recursion, which led to not enough spawners being
added to the list, and not having enough spawners to assign the patterns.

Function SpawnerFactory had to be heavily reworked. Now it works correclty,
but is also simpler, and makes more sense.

Function ParseSpawnPattern is changed to return a dictionary instead of an array:
1) It is more convenient for being "piped" into SpawnerFactory.
2) It doesn't make sense converting the data from a dictionary to an array
before returning, and then converting the array back to a dictionary.

Signed-off-by: AnimaRain <ioannis_spyr@yahoo.gr>